refactor: shuffle around a type
that way we can use it in this new handler?
